+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 Changelog
 ---------
 
+**[0.4]**
+   - Fixed bug in evaluation of precedence when an option is defined in
+     more than one source.
+   - Fixed bug in defaults where a default of `False` would not be returned.
+
 **[0.3]**
    - Added `default` argument to Option.  Now Options can be declared with
      a default value to be returned if the option value is not found in